Replace deprecated np.int and np.float types

+++ pysptools/abundance_maps/amaps_int.py
@@ -190,7 +190,7 @@ def _document(cls):
 
 def _compress(vec, mask):
     n = np.sum(mask)
-    cmp = np.ndarray((n, vec.shape[1]), dtype=np.float)
+    cmp = np.ndarray((n, vec.shape[1]), dtype=np.float32)
     i = 0
     for j in range(mask.shape[0]):
         if mask[j] == 1:
@@ -200,7 +200,7 @@ def _compress(vec, mask):
 
 
 def _expand(amap, mask, l, q):
-    exp = np.zeros((l,q), dtype=np.float)
+    exp = np.zeros((l,q), dtype=np.float32)
     i = 0
     for j in range(mask.shape[0]):
         if mask[j] == 1:
